+++ lib/td_cache/event_stream.ex
@@ -4,17 +4,14 @@ defmodule TdCache.EventStream do
   """
 
   def child_spec(config) do
-    consumer = config[:consumer]
-    group = config[:group]
+    consumer_id = config[:consumer_id]
+    consumer_group = config[:consumer_group]
 
     children =
       config[:streams]
       |> Enum.with_index()
-      |> Enum.map(fn {stream, i} ->
-        Supervisor.child_spec(
-          {TdCache.EventStream.Consumer, stream: stream, group: group, consumer: consumer},
-          id: {EventStream, i}
-        )
+      |> Enum.flat_map(fn {stream_config, i} ->
+        stream_workers(stream_config, consumer_group, consumer_id, i)
       end)
 
     # Spec for the supervisor that will supervise the EventStream consumers.
@@ -23,5 +20,21 @@ defmodule TdCache.EventStream do
       type: :supervisor,
       start: {Supervisor, :start_link, [children, [strategy: :one_for_one]]}
     }
+  end
+
+  defp stream_workers(stream_config, consumer_group, consumer_id, i) do
+    stream = stream_config[:key]
+    consumer = stream_config[:consumer]
+
+    [
+      Supervisor.child_spec(
+        {TdCache.EventStream.Consumer,
+         stream: stream,
+         consumer_group: consumer_group,
+         consumer_id: consumer_id,
+         consumer: consumer},
+        id: {EventStream.Consumer, i}
+      )
+    ]
   end
 end