+++ lib/td_cache/event_stream/processor.ex
@@ -1,0 +1,5 @@
+defmodule TdCache.EventStream.Processor do
+  @moduledoc """
+  Defines the contract to process events read from a `TdCache.EventStream`
+  """
+end