+++ test/td_cache/event_stream/consumer_test.exs
@@ -1,55 +1,52 @@
 defmodule TdCache.EventStream.ConsumerTest do
   use ExUnit.Case
   alias TdCache.EventStream.Consumer
+  alias TdCache.EventStream.TestConsumer
   alias TdCache.Redix, as: Redis
   doctest TdCache.EventStream.Consumer
 
-  setup_all do
+  setup do
     stream = "test:events"
-    group = "test_consumer_group"
 
-    options = [
+    config = [
+      consumer_group: "test_group",
+      consumer_id: "test_consumer",
       stream: stream,
-      group: group,
-      consumer: "default",
+      consumer: TestConsumer,
       parent: self()
     ]
 
-    {:ok, _} = Redis.command(["DEL", stream])
-    :ok = start_consumer(options)
-    {:ok, stream: stream, group: group}
+    on_exit(fn -> Redis.command!(["DEL", stream]) end)
+
+    {:ok, _pid} = TestConsumer.start_link(parent: self())
+    :ok = start_consumer(config)
+    {:ok, stream: stream}
   end
 
-  defp start_consumer(options) do
-    {:ok, _pid} = Consumer.start_link(options)
+  defp start_consumer(config) do
+    {:ok, _pid} = Consumer.start_link(config)
 
     receive do
       :started -> :ok
     after
-      1_000 -> :timeout
+      2_000 -> :timeout
     end
   end
 
-  describe "EventStream" do
-    test "creates a consumer group on startup", context do
-      stream = context[:stream]
-      group = context[:group]
-      {:ok, groups} = Redis.command(["XINFO", "GROUPS", stream])
-      assert Enum.any?(groups, fn [_, name, _, _, _, _, _, _] -> name == group end)
-    end
-
-    test "consumes and acknowledges an event which is on the stream", context do
+  describe "EventStream Consumer" do
+    test "consumes events on the stream", context do
       stream = context[:stream]
       {:ok, event_id} = Redis.command(["XADD", stream, "*", "foo", "bar"])
+      {:consumed, events} = consume_events()
+      assert Enum.any?(events, &(&1.id == event_id))
+    end
+  end
 
-      {:ok, events} = Consumer.read(stream)
-      assert Enum.count(events) >= 1
-      assert Enum.any?(events, &(Map.get(&1, :foo) == "bar"))
-      event_ids = events |> Enum.map(& &1.id)
-      assert Enum.member?(event_ids, event_id)
-
-      {:ok, count} = Consumer.ack(stream, event_ids)
-      assert count == Enum.count(event_ids)
+  defp consume_events do
+    receive do
+      m -> m
+    after
+      5_000 -> :timeout
     end
   end
 end