+++ test/td_cache/structure_cache_test.exs
@@ -1,11 +1,11 @@
 defmodule TdCache.StructureCacheTest do
   use ExUnit.Case
   alias TdCache.StructureCache
+  alias TdCache.SystemCache
   doctest TdCache.StructureCache
 
   setup do
     system = %{id: :rand.uniform(100_000_000), external_id: "foo", name: "bar"}
-    metadata = %{foo: "bar"}
 
     structure = %{
       id: :rand.uniform(100_000_000),
@@ -13,9 +13,13 @@ defmodule TdCache.StructureCacheTest do
       group: "group",
       type: "type",
       path: ["foo", "bar"],
-      metadata: metadata,
       system: system
     }
+
+    on_exit(fn ->
+      StructureCache.delete(structure.id)
+      SystemCache.delete(system.id)
+    end)
 
     {:ok, structure: structure}
   end
@@ -29,14 +33,13 @@ defmodule TdCache.StructureCacheTest do
       structure = context[:structure]
       {:ok, _} = StructureCache.put(structure)
       {:ok, s} = StructureCache.get(structure.id)
-      assert Map.get(s, :system) == Map.drop(structure.system, [:id])
-      assert Map.drop(s, [:id, :system]) == Map.drop(structure, [:id, :system])
+      assert s == structure
     end
 
     test "deletes an entry in redis", context do
       structure = context[:structure]
-      assert {:ok, ["OK", 0, 2, "OK", "OK"]} = StructureCache.put(structure)
-      assert {:ok, 3} = StructureCache.delete(structure.id)
+      assert {:ok, ["OK", 0, 2, "OK"]} = StructureCache.put(structure)
+      assert {:ok, 2} = StructureCache.delete(structure.id)
       assert {:ok, nil} = StructureCache.get(structure.id)
     end
   end