+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [3.0.0] 2019-06-14
+
+### Added
+
+- Cache for fields, structures, systems and links
+- Redis stream consumer
+- Use connection pool for Redis
+
 ## [2.21.1] 2019-06-06
 
 ### Changed